pyproject.toml
[project]
name = "gv-auto"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Pavel", email = "60060559+pyrogn@users.noreply.github.com" }
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "seleniumbase>=4.27.5",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gv_auto"]
